infrastructure/aws/cdk/requirements.txt
aws-cdk-lib>=2.234.1,<3.0.0
constructs>=10.0.0,<11.0.0
cdk-nag>=2.27.0  # For security guardrails